+++ src/main/java/com/ujcms/cms/core/mapper/SiteMapper.java
@@ -31,13 +31,11 @@ public interface SiteMapper extends TreeEntityMapper<Site> {
      * 根据查询条件获取列表
      *
      * @param queryInfo          查询条件
-     * @param customsCondition   自定义字段查询条件
      * @param isQueryHasChildren 是否查询包含子站点
      * @param fullOrgId          组织ID。属于该组织的上级或下级的站点都符合条件
      * @return 数据列表
      */
     List<Site> selectAll(@Nullable @Param("queryInfo") QueryInfo queryInfo,
-                         @Nullable @Param("customsCondition") List<QueryInfo.WhereCondition> customsCondition,
                          @Param("isQueryHasChildren") boolean isQueryHasChildren,
                          @Nullable @Param("fullOrgId") Long fullOrgId);