+++ src/main/java/com/ujcms/cms/core/service/SiteService.java
@@ -7,7 +7,6 @@ import com.ujcms.cms.core.domain.Site;
 import com.ujcms.cms.core.domain.generated.GeneratedSite;
 import com.ujcms.cms.core.mapper.SiteMapper;
 import com.ujcms.cms.core.service.args.SiteArgs;
-import com.ujcms.commons.query.CustomFieldQuery;
 import com.ujcms.commons.query.QueryInfo;
 import com.ujcms.commons.query.QueryParser;
 import org.owasp.html.PolicyFactory;
@@ -79,8 +78,7 @@ public class SiteService {
 
     public List<Site> selectList(SiteArgs args) {
         QueryInfo queryInfo = QueryParser.parse(args.getQueryMap(), GeneratedSite.TABLE_NAME, "order,id");
-        List<QueryInfo.WhereCondition> customsCondition = CustomFieldQuery.parse(args.getCustomsQueryMap());
-        return mapper.selectAll(queryInfo, customsCondition, args.isQueryHasChildren(), args.getFullOrgId());
+        return mapper.selectAll(queryInfo, args.isQueryHasChildren(), args.getFullOrgId());
     }
 
     public List<Site> selectList(SiteArgs args, int offset, int limit) {